requirements-pycamp.txt
asgiref==3.8.1
astroid==2.3.3
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backports.csv==1.0.7
cairocffi==1.7.1
CairoSVG==2.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
coverage==7.5.3
cryptography==42.0.8
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-allauth==0.63.3
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-autofixture @ git+https://github.com/josx/django-autofixture@48c400e20fe54f7af5d254b57780244787a1953f
django-ckeditor==6.7.1
django-configurations==2.5.1
django-debug-toolbar==3.5.0
django-email-extras==0.3.4
django-extensions==3.2.3
django-filter==21.1
django-formset-js==0.5.0
django-image-cropping==1.7
django-import-export==4.0.9
django-jquery-js==3.1.1
django-js-asset==2.2.0
django-jsonfield-backport==1.0.5
django-queryset-csv==1.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-tempus-dominus==5.1.2.17
django-vote==2.4
django-webpack-loader==0.5.0
djangorestframework==3.15.1
djangorestframework-queryfields==1.0.0
docopt==0.6.2
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
exceptiongroup==1.2.1
execnet==2.1.1
executing==2.0.1
future==0.15.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipython==8.18.1
isort==4.3.21
jedi==0.19.1
lazy-object-proxy==1.4.3
lxml==5.2.2
matplotlib-inline==0.1.7
mccabe==0.6.1
more-itertools==10.3.0
msgpack-python==0.5.6
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.4
packaging==24.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.0.1
pluggy==0.13.1
posix-ipc==1.1.1
prompt_toolkit==3.0.47
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.22
Pygments==2.18.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.8.2
pypng==0.0.18
PyQRCode==1.2.1
pytest==5.3.0
pytest-cov==2.8.1
pytest-django==3.7.0
pytest-forked==1.6.0
pytest-gitignore==1.3
pytest-lazy-fixture==0.6.2
pytest-mock==1.12.1
pytest-sugar==0.9.2
pytest-watch==4.2.0
pytest-xdist==1.30.0
python-gnupg==0.5.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
raven==6.9.0
redis==2.10.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sphinx-me==0.3
sqlparse==0.5.0
stack-data==0.6.3
svglue==0.3.0
tablib==3.5.0
termcolor==2.4.0
tinycss2==1.3.0
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
unicodecsv==0.14.1
Unidecode==1.3.8
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.11.2
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.4.post2
jwt==1.3.1
